can now close socket

+++ services/master/collector/message.go
@@ -1,6 +1,6 @@
 package collector
 
 type holeMessage struct {
-	FrameIdx     int64 `json:"frameIdx"`
-	PlayerNumber int   `json:"playerNumber"`
+	FrameIdx     int `json:"frameIdx"`
+	PlayerNumber int `json:"playerNumber"`
 }
